Add switching between attribute types. Add month rendering. Fix lack of precision on averages.

+++ CoachBot.Domain/Model/PlayerPerformanceSnapshot.cs
@@ -16,11 +16,11 @@ namespace CoachBot.Domain.Model
 
         public int Year { get; set; }
 
-        public int AverageGoals { get; set; }
+        public double AverageGoals { get; set; }
 
-        public int AverageAssists { get; set; }
+        public double AverageAssists { get; set; }
 
-        public int AverageGoalsConceded { get; set; }
+        public double AverageGoalsConceded { get; set; }
 
         public int CleanSheets { get; set; }
